fix: circular dependency on VoiceMessageInput types (#423)
