+++ src/smart-components/Channel/components/MessageInput/VoiceMessageInputWrapper.tsx
@@ -7,7 +7,8 @@ import { useVoicePlayer } from '../../../../hooks/VoicePlayer/useVoicePlayer';
 import { useVoiceRecorder, VoiceRecorderStatus } from '../../../../hooks/VoiceRecorder/useVoiceRecorder';
 import { isDisabledBecauseFrozen, isDisabledBecauseMuted } from '../../context/utils';
 
-import { VoiceMessageInput, VoiceMessageInputStatus } from '../../../../ui/VoiceMessageInput';
+import { VoiceMessageInput } from '../../../../ui/VoiceMessageInput';
+import { VoiceMessageInputStatus } from '../../../../ui/VoiceMessageInput/types';
 import Modal from '../../../../ui/Modal';
 import Button, { ButtonSizes, ButtonTypes } from '../../../../ui/Button';
 import useSendbirdStateContext from '../../../../hooks/useSendbirdStateContext';
